Changed MintFrameTimer to count in seconds, update animation related examples, updated the phys test to have feel resembling GraveyardShift

+++ src/mintSDL/maths/phys.h
@@ -20,7 +20,7 @@ struct MintPhys {
 	MintDoublePoint maxVelocity;
 };
 
-void mint_PhysUpdate(MintPhys* phys, float elapsed);
+void mint_PhysUpdate(MintPhys* phys, double elapsed);
 double mint_PhysComputeVelocity(double velocity, double accel, double drag, double max, double elapsed);
 
 #endif
